LungCancerDetection/requirements.txt
flask==2.3.3
flask-cors==4.0.0
tensorflow==2.15.0
pillow==10.2.0
numpy==1.26.3
pymongo==4.6.1
werkzeug==2.3.7
python-dotenv==1.0.0 